abstract the logic of "useSignTxWithDevice" into an agnostic hook
