+++ src/hooks/transactions/useSignMultipleTransactions.tsx
@@ -1,0 +1,224 @@
+import { useEffect, useState } from 'react';
+import { Transaction } from '@elrondnetwork/erdjs';
+import { getScamAddressData } from 'apiCalls';
+import { useParseMultiEsdtTransferData } from 'services/transactions/hooks/useParseMultiEsdtTransferData';
+import { ActiveLedgerTransactionType, MultiSignTxType } from 'types';
+import { LoginMethodsEnum } from 'types/enums';
+import { getIsProviderEqualTo, isTokenTransfer } from 'utils';
+import { getLedgerErrorCodes } from 'utils/internal';
+
+export interface UseSignMultipleTransactionsPropsType {
+  egldLabel: string;
+  address: string;
+  verifyReceiverScam?: boolean;
+  transactionsToSign?: Transaction[];
+  onCancel: () => void;
+  onSignTransaction: (transaction: Transaction) => Promise<Transaction>;
+  onTransactionsSignSuccess: (transactions: Transaction[]) => void;
+  onTransactionsSignError: (errorMessage: string) => void;
+}
+
+interface VerifiedAddressesType {
+  [address: string]: { type: string; info: string };
+}
+let verifiedAddresses: VerifiedAddressesType = {};
+
+type DeviceSignedTransactions = Record<number, Transaction>;
+
+export interface UseSignMultipleTransactionsReturnType {
+  allTransactions: MultiSignTxType[];
+  onSignTransaction: () => void;
+  onNext: () => void;
+  onPrev: () => void;
+  onAbort: () => void;
+  waitingForDevice: boolean;
+  isLastTransaction: boolean;
+  currentStep: number;
+  signedTransactions?: DeviceSignedTransactions;
+  currentTransaction: ActiveLedgerTransactionType | null;
+}
+
+export function useSignMultipleTransactions({
+  transactionsToSign,
+  egldLabel,
+  address,
+  verifyReceiverScam = true,
+  onCancel,
+  onSignTransaction,
+  onTransactionsSignError,
+  onTransactionsSignSuccess
+}: UseSignMultipleTransactionsPropsType): UseSignMultipleTransactionsReturnType {
+  const [currentStep, setCurrentStep] = useState(0);
+  const [signedTransactions, setSignedTransactions] =
+    useState<DeviceSignedTransactions>();
+  const [currentTransaction, setCurrentTransaction] =
+    useState<ActiveLedgerTransactionType | null>(null);
+
+  const [waitingForDevice, setWaitingForDevice] = useState(false);
+  const isLedger = getIsProviderEqualTo(LoginMethodsEnum.ledger);
+
+  const { getTxInfoByDataField, allTransactions } =
+    useParseMultiEsdtTransferData({ transactions: transactionsToSign });
+
+  const isLastTransaction = currentStep === allTransactions.length - 1;
+
+  useEffect(() => {
+    extractTransactionsInfo();
+  }, [currentStep, allTransactions]);
+
+  async function extractTransactionsInfo() {
+    const tx = allTransactions[currentStep];
+    if (tx == null) {
+      return;
+    }
+    const { transaction, multiTxData } = tx;
+    const dataField = transaction.getData().toString();
+    const transactionTokenInfo = getTxInfoByDataField(
+      transaction.getData().toString(),
+      multiTxData
+    );
+    const { tokenId } = transactionTokenInfo;
+    const receiver = transaction.getReceiver().toString();
+    const notSender = address !== receiver;
+    const verified = receiver in verifiedAddresses;
+
+    if (notSender && !verified && verifyReceiverScam) {
+      const data = await getScamAddressData(receiver);
+      verifiedAddresses = {
+        ...verifiedAddresses,
+        ...(data.scamInfo ? { [receiver]: data.scamInfo } : {})
+      };
+    }
+
+    const isTokenTransaction = Boolean(
+      tokenId && isTokenTransfer({ tokenId, erdLabel: egldLabel })
+    );
+
+    setCurrentTransaction({
+      transaction,
+      receiverScamInfo: verifiedAddresses[receiver]?.info || null,
+      transactionTokenInfo,
+      isTokenTransaction,
+      dataField
+    });
+  }
+
+  function reset() {
+    setCurrentStep(0);
+    setSignedTransactions(undefined);
+    setWaitingForDevice(false);
+  }
+
+  async function sign() {
+    try {
+      if (currentTransaction == null) {
+        return;
+      }
+
+      setWaitingForDevice(isLedger);
+
+      const signedTx = await onSignTransaction(currentTransaction.transaction);
+      const newSignedTx = { [currentStep]: signedTx };
+      const newSignedTransactions = signedTransactions
+        ? { ...signedTransactions, ...newSignedTx }
+        : newSignedTx;
+      setSignedTransactions(newSignedTransactions);
+      if (!isLastTransaction) {
+        setCurrentStep((exising) => exising + 1);
+        setWaitingForDevice(false);
+      } else if (newSignedTransactions) {
+        onTransactionsSignSuccess(Object.values(newSignedTransactions));
+        reset();
+      }
+    } catch (err) {
+      console.error(err, 'sign error');
+      const { message } = err as any;
+      const errorMessage = isLedger
+        ? getLedgerErrorCodes(err).errorMessage
+        : null;
+
+      reset();
+      onTransactionsSignError(errorMessage ?? message);
+    }
+  }
+
+  function signTx() {
+    try {
+      if (currentTransaction == null) {
+        return;
+      }
+      const signature = currentTransaction.transaction.getSignature();
+      if (signature.hex()) {
+        if (!isLastTransaction) {
+          setCurrentStep((exising) => exising + 1);
+        }
+      } else {
+        // currently code doesn't reach here because getSignature throws error if none is found
+        sign();
+      }
+    } catch {
+      // the only way to check if tx has signature is with try catch
+      sign();
+    }
+  }
+
+  const isFirst = currentStep === 0;
+
+  function handleAbort() {
+    if (isFirst) {
+      onCancel();
+    } else {
+      setCurrentStep((existing) => existing - 1);
+    }
+  }
+
+  const continueWithoutSigning =
+    currentTransaction?.transactionTokenInfo?.type &&
+    currentTransaction?.transactionTokenInfo?.multiTxData &&
+    !currentTransaction?.dataField.endsWith(
+      currentTransaction?.transactionTokenInfo?.multiTxData
+    );
+
+  function handleSignTransaction() {
+    if (continueWithoutSigning) {
+      setCurrentStep((exising) => exising + 1);
+    } else {
+      signTx();
+    }
+  }
+
+  function onNext() {
+    setCurrentStep((current) => {
+      const nextStep = current + 1;
+      if (nextStep > allTransactions?.length) {
+        return current;
+      }
+      return nextStep;
+    });
+  }
+
+  function onPrev() {
+    setCurrentStep((current) => {
+      const nextStep = current - 1;
+      if (nextStep < 0) {
+        return current;
+      }
+      return nextStep;
+    });
+  }
+
+  return {
+    allTransactions,
+    onSignTransaction: handleSignTransaction,
+    onNext,
+    onPrev,
+    waitingForDevice,
+    onAbort: handleAbort,
+    isLastTransaction,
+    currentStep,
+    signedTransactions,
+    currentTransaction
+  };
+}
+
+export default useSignMultipleTransactions;
